Crate matrix_sdk_appservice

Matrix Application Service library

The appservice crate aims to provide a batteries-included experience by being a thin wrapper around the [matrix_sdk]. That means that we

ship with functionality to configure your webserver crate or simply run the webserver for you

receive and validate requests from the homeserver correctly

allow calling the homeserver with proper virtual user identity assertion

have consistent room state by leveraging matrix-sdk’s state store

provide E2EE support by leveraging matrix-sdk’s crypto store

Status

The crate is in an experimental state. Follow matrix-org/matrix-rust-sdk#228 for progress.

Registration

The crate relies on the appservice registration being always in sync with the actual registration used by the homeserver. That’s because it’s required for the access tokens and because membership states for virtual users are determined based on the registered namespaces.

Quickstart

use matrix_sdk_appservice::{
    ruma::events::room::member::SyncRoomMemberEvent, AppService,
    AppServiceRegistration,
};

let homeserver_url = "http://127.0.0.1:8008";
let server_name = "localhost";
let registration = AppServiceRegistration::try_from_yaml_file(
    "./tests/registration.yaml",
)?;

let mut appservice = AppService::builder(
    homeserver_url.try_into()?,
    server_name.try_into()?,
    registration,
)
.build()
.await?;
appservice.virtual_user(None).await?.add_event_handler(
    |_ev: SyncRoomMemberEvent| async {
        // do stuff
    },
);

let (host, port) = appservice.registration().get_host_and_port()?;
appservice.run(host, port).await?;

Check the examples directory for fully working examples.

Re-exports

pub use matrix_sdk;

pub use matrix_sdk::ruma;

pub use registration::AppServiceRegistration;

pub use virtual_user::VirtualUserBuilder;

Modules

event_handler

registration

AppService Registration.

virtual_user

Virtual users.

Structs

AppService

AppService

AppServiceBuilder

Builder for an AppService

AppServiceRouter

Enums

Error

Type Definitions

Result