Crate matrix_sdk_appservice::api

(De)serializable types for the Matrix Application Service API. These types can be shared by application service and server code.

Modules

event
Endpoint for sending events.

query
Endpoints for querying user IDs and room aliases

thirdparty
Endpoints for third party lookups

Structs

Namespace
A namespace defined by an application service.

Namespaces
Namespaces defined by an application service.

Registration
Information required in the registration yaml file that a homeserver needs.

RegistrationInit
Initial set of fields of Registration.