Struct matrix_sdk_appservice::AppService

pub struct AppService { /* private fields */ }

AppService

Implementations

impl AppService

pub fn builder( homeserver_url: Url, server_name: OwnedServerName, registration: AppServiceRegistration ) -> AppServiceBuilder

Create a new AppServiceBuilder.

pub async fn user(&self, localpart: Option<&str>) -> Result<Client>

Create an appservice user client.

Will create and return a client that’s configured to assert the identity on outgoing homeserver requests that need authentication.

This method is a singleton that saves the client internally for re-use based on the localpart. The cached client can be retrieved by calling this method again.

Note that if you want to do actions like joining rooms with a user it needs to be registered first. register_user() can be used for that purpose.

Arguments

localpart - Used for constructing the user accordingly. If None is given it uses the sender_localpart from the registration.

pub async fn user_with_client_builder( &self, localpart: Option<&str>, builder: ClientBuilder ) -> Result<Client>

Same as user() but with the ability to pass in a [ClientBuilder].

Since this method is a singleton follow-up calls with different [ClientBuilder]s will be ignored.

pub fn user_builder<'a>(&'a self, localpart: &'a str) -> UserBuilder<'a>

Create a new appservice user builder for the given localpart.

pub fn users(&self) -> Arc<DashMap<String, Client>>

Get the map containing all constructed appservice user clients.

pub async fn register_user_query( &self, handler: Box<dyn FnMut(AppService, Request) -> Pin<Box<dyn Future<Output = bool> + Send + 'static>> + Send + Sync + 'static> )

Register a responder for queries about the existence of a user with a given mxid.

See GET /_matrix/app/v1/users/{userId}.

Example

appservice.register_user_query(Box::new(|appservice, req| {
    Box::pin(async move {
        println!("Got request for {}", req.user_id);
        true
    })
}));

pub async fn register_room_query( &self, handler: Box<dyn FnMut(AppService, Request) -> Pin<Box<dyn Future<Output = bool> + Send + 'static>> + Send + Sync + 'static> )

Register a responder for queries about the existence of a room with the given alias.

See GET /_matrix/app/v1/rooms/{roomAlias}.

Example

appservice.register_room_query(Box::new(|appservice, req| {
    Box::pin(async move {
        println!("Got request for {}", req.room_alias);
        true
    })
}));

pub async fn register_user( &self, localpart: &str, device_id: Option<&DeviceId> ) -> Result<()>

Register an appservice user by sending a [register::v3::Request] to the homeserver.

Arguments

localpart - The localpart of the user to register. Must be covered by the namespaces in the registration in order to succeed.

Returns

This function may return a UIAA response, which should be checked for with Error::as_uiaa_response().

pub fn registration(&self) -> &AppServiceRegistration

Get the AppServiceRegistration.

pub fn compare_hs_token(&self, hs_token: impl AsRef<str>) -> bool

Compare the given hs_token against the registration’s hs_token.

Returns true if the tokens match, false otherwise.

pub fn user_id_is_in_namespace(&self, user_id: impl AsRef<str>) -> bool

Check if given user_id is in any of the AppServiceRegistration’s users namespaces.

pub fn service<B>(&self) -> AppServiceRouter<B>where B: HttpBody + Send + 'static, B::Data: Send, B::Error: Into<BoxError>,

Returns a [Service][tower::Service] that processes appservice requests.

pub async fn run( &self, host: impl Into<String>, port: impl Into<u16> ) -> Result<()>

Convenience method that runs an http server.

This is a blocking call that tries to listen on the provided host and port.

pub fn set_default_request_config( &mut self, request_config: Option<RequestConfig> ) -> Result<()>

Set the default RequestConfig