Crate matrix_sdk_appservice

Matrix Application Service library

The appservice crate aims to provide a batteries-included experience by being a thin wrapper around the matrix_sdk. That means that we

ship with functionality to configure your webserver crate or simply run the webserver for you

receive and validate requests from the homeserver correctly

allow calling the homeserver with proper virtual user identity assertion

have consistent room state by leveraging matrix-sdk’s state store

provide E2EE support by leveraging matrix-sdk’s crypto store

Status

The crate is in an experimental state. Follow matrix-org/matrix-rust-sdk#228 for progress.

Quickstart

use matrix_sdk_appservice::{
    ruma::events::room::member::SyncRoomMemberEvent,
    AppService, AppServiceRegistration
};

let homeserver_url = "http://127.0.0.1:8008";
let server_name = "localhost";
let registration = AppServiceRegistration::try_from_yaml_str(
    r"
        id: appservice
        url: http://127.0.0.1:9009
        as_token: as_token
        hs_token: hs_token
        sender_localpart: _appservice
        namespaces:
          users:
          - exclusive: true
            regex: '@_appservice_.*'
    ")?;

let mut appservice = AppService::new(homeserver_url, server_name, registration).await?;
appservice.register_event_handler(|_ev: SyncRoomMemberEvent| async {
    // do stuff
});

let (host, port) = appservice.registration().get_host_and_port()?;
appservice.run(host, port).await?;

Check the examples directory for fully working examples.

Re-exports

pub use matrix_sdk;

pub use matrix_sdk::ruma;

Modules

event_handler

Structs

AppService

AppService

AppServiceRegistration

AppService Registration

Enums

Error

Type Definitions

Host

MainUser

The localpart of the user associated with the application service via sender_localpart in AppServiceRegistration.

Port

Result

VirtualUser

The application service may specify the virtual user to act as through use of a user_id query string parameter on the request. The user specified in the query string must be covered by one of the AppServiceRegistration’s users namespaces.