Crate matrix_sdk_appservice

Matrix Application Service library

The appservice crate aims to provide a batteries-included experience. That means that we

ship with functionality to configure your webserver crate or simply run the webserver for you

receive and validate requests from the homeserver correctly

allow calling the homeserver with proper virtual user identity assertion

have the goal to have a consistent room state available by leveraging the stores that the matrix-sdk provides

Quickstart

use matrix_sdk_appservice::{Appservice, AppserviceRegistration};

let homeserver_url = "http://127.0.0.1:8008";
let server_name = "localhost";
let registration = AppserviceRegistration::try_from_yaml_str(
    r"
        id: appservice
        url: http://127.0.0.1:9009
        as_token: as_token
        hs_token: hs_token
        sender_localpart: _appservice
        namespaces:
          users:
          - exclusive: true
            regex: '@_appservice_.*'
    ")
    .unwrap();

let appservice = Appservice::new(homeserver_url, server_name, registration).await.unwrap();
// set event handler with `appservice.client().set_event_handler()` here
let (host, port) = appservice.get_host_and_port_from_registration().unwrap();
appservice.run(host, port).await.unwrap();

Modules

api
(De)serializable types for the Matrix Application Service API. These types can be shared by application service and server code.

Structs

Appservice
Appservice

AppserviceRegistration
Appservice Registration

Enums

Error

Type Definitions

Host

Port

Result