Struct matrix_sdk_appservice::AppService

pub struct AppService { /* private fields */ }

AppService

Implementations

impl AppService

pub async fn new(
homeserver_url: impl TryInto<Url, Error = ParseError>,
server_name: impl TryInto<OwnedServerName, Error = IdParseError>,
registration: AppServiceRegistration
) -> Result<Self>

Create new AppService

Also creates and caches a [Client] for the MainUser. A default [ClientBuilder] is used, if you want to customize it use with_client_builder() instead.

Arguments

homeserver_url - The homeserver that the client should connect to.

server_name - The server name to use when constructing user ids from the localpart.

registration - The AppService Registration to use when interacting with the homeserver.

pub async fn with_client_builder(
homeserver_url: impl TryInto<Url, Error = ParseError>,
server_name: impl TryInto<OwnedServerName, Error = IdParseError>,
registration: AppServiceRegistration,
builder: ClientBuilder
) -> Result<Self>

Same as new() but lets you provide a [ClientBuilder] for the [Client]

pub async fn virtual_user_client(
&self,
localpart: impl AsRef<str>
) -> Result<Client>

Create a [Client] for the given VirtualUser’s localpart

Will create and return a [Client] that’s configured to assert the identity on all outgoing homeserver requests if localpart is given.

This method is a singleton that saves the client internally for re-use based on the localpart. The cached [Client] can be retrieved either by calling this method again or by calling get_cached_client() which is non-async convenience wrapper.

Note that if you want to do actions like joining rooms with a virtual user it needs to be registered first. Self::register_virtual_user() can be used for that purpose.

Arguments

localpart - The localpart of the user we want assert our identity to

pub async fn virtual_user_client_with_client_builder(
&self,
localpart: impl AsRef<str>,
builder: ClientBuilder
) -> Result<Client>

Same as virtual_user_client() but with the ability to pass in a [ClientBuilder]

Since this method is a singleton follow-up calls with different [ClientBuilder]s will be ignored.

pub fn get_cached_client(&self, localpart: Option<&str>) -> Result<Client>

Get cached [Client]

Will return the client for the given localpart if previously constructed with virtual_user_client() or [virtual_user_client_with_config()][Self:: virtual_user_client_with_client_builder].

If no localpart is given it assumes the MainUser’s localpart. If no client for localpart is found it will return an Error.

pub async fn register_event_handler<Ev, Ctx, H>(
&self,
handler: H
) -> Result<&Self> where
Ev: SyncEvent + DeserializeOwned + Send + 'static,
H: EventHandler<Ev, Ctx>,
<H::Future as Future>::Output: EventHandlerResult,

Convenience wrapper around [Client::register_event_handler()] that attaches the event handler to the MainUser’s [Client]

Note that the event handler in the AppService context only triggers join room timeline events, so no state events or events from the invite, knock or leave scope. The rationale behind that is that incoming AppService transactions from the homeserver are not necessarily bound to a specific user but can cover a multitude of namespaces, and as such the AppService basically only “observes joined rooms”. Also currently homeservers only push PDUs to appservices, no EDUs. There’s the open MSC2409 regarding supporting EDUs in the future, though it seems to be planned to put EDUs into a different JSON key than events to stay backwards compatible.

pub fn register_event_handler_context<T>(&self, ctx: T) -> Result<&Self> where
T: Clone + Send + Sync + 'static,

Convenience wrapper around [Client::register_event_handler_context] attaches the event handler context to the MainUser’s [Client].

pub async fn register_user_query(
&self,
handler: Box<dyn FnMut(AppService, IncomingRequest) -> Pin<Box<dyn Future<Output = bool> + Send + 'static>> + Send + Sync + 'static>
)

Register a responder for queries about the existence of a user with a given mxid.

See GET /_matrix/app/v1/users/{userId}.

Example

appservice.register_user_query(Box::new(|appservice, req| Box::pin(async move {
    println!("Got request for {}", req.user_id);
    true
})));

pub async fn register_room_query(
&self,
handler: Box<dyn FnMut(AppService, IncomingRequest) -> Pin<Box<dyn Future<Output = bool> + Send + 'static>> + Send + Sync + 'static>
)

Register a responder for queries about the existence of a room with the given alias.

See GET /_matrix/app/v1/rooms/{roomAlias}.

Example

appservice.register_room_query(Box::new(|appservice, req| Box::pin(async move {
    println!("Got request for {}", req.room_alias);
    true
})));

pub async fn register_virtual_user(
&self,
localpart: impl AsRef<str>
) -> Result<()>

Register a virtual user by sending a [register::v3::Request] to the homeserver

Arguments

localpart - The localpart of the user to register. Must be covered by the namespaces in the [Registration] in order to succeed.

Returns

This function may return a UIAA response, which should be checked for with Error::uiaa_response().

pub fn registration(&self) -> &AppServiceRegistration

Get the AppService registration

pub fn compare_hs_token(&self, hs_token: impl AsRef<str>) -> bool

Compare the given hs_token against registration.hs_token

Returns true if the tokens match, false otherwise.

pub fn user_id_is_in_namespace(&self, user_id: impl AsRef<str>) -> Result<bool>

Check if given user_id is in any of the AppServiceRegistration’s users namespaces

pub fn warp_filter(&self) -> BoxedFilter<(impl Reply,)>

Returns a warp::Filter to be used as warp::serve() route

Note that if you handle any of the application-service-specific routes, including the legacy routes, you will break the appservice functionality.

pub async fn run(
&self,
host: impl Into<String>,
port: impl Into<u16>
) -> Result<()>

Convenience method that runs an http server depending on the selected server feature

This is a blocking call that tries to listen on the provided host and port